feat: eigene Seite für letzte Aufsichten statt Popover
- Verlauf als eigene Route /session-history mit großen Zeilen und
  Pagination (5 pro Seite) statt scrollbarem Popover unten links
- Zugang als Sekundär-Button unter der Aufsicht-starten-Karte
- Wording stellt klar: Einträge sind beendete Aufsichten, Auswahl
  startet eine neue Aufsicht mit derselben Kombination
